rename createBranch file to createReleaseBump
